requirements-dev.txt
jupyter
sodapy
python-dotenv
GitPython